chore(api): change select network route path and improve redis logic
